pyproject.toml
[tool.poetry]
name = "wow-classic-fishing-bot"
version = "0.1.0"
description = ""
authors = ["Adam Yala <adamkyala@gmail.com>"]
readme = "README.md"
packages = [{include = "wow_classic_fishing_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.5.0"
pyautogui = "^0.9.53"
numpy = "^1.24.3"
opencv-python = "^4.7.0.72"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
